Refactor namespaces and update DictionaryDBContext
- Changed namespaces from `AnagramAPI` to `RoachagramAPI` for `AnagramController`, `AIModel`, and `Models` classes.
- Added `using` directive for `Roachagram.API.Models` in `AnagramController.cs` and `Startup.cs`.
- Modified `DictionaryDBContext` constructor in `Models.cs` to improve database context configuration.
- Updated comments in `Models.cs` for clearer documentation.
