+++ Roachagram.API/Models/AIModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace AnagramAPI.Models
+namespace RoachagramAPI.Models
 {
     /// <summary>
     ///
